workers/requirements.txt
jinja2==3.1.2
orjson==3.7.12
aiosmtplib-1.1.7